+++ doc/version-history.rst
@@ -5,6 +5,7 @@ Version History
 v1.10.0
 =======
 
+* getStringStatus function
 * handleMissingReply, MissingReply exception
 * IRQTimeout handling